Move include ptah logic to new class
